Map incomes and expenses as classes in index.py
Then uses marshmallow to serialize and deseralize objects.

Redefines the incomes variable into a list of Expenses and Incomes, now
called transactions. Refactored both existing endpoints accordingly.

For the endpont used to retrieve incomes, deines an instance of
IncomeSchema to produce JASON representation of incomes. Uses filter to
extract incomes only from the transactions list. At the end, sends the
array of JASON incomes back to users.

For the endpoint used to accept new incomes, add IncomeSchema to load an
instance of Income based on the JSON data sent by the user. Add the new
Income to the transactions list.

There are0 two new endpoints to handle expenses, which are similar to
the respective income endpoints.